updating/adding - admin CRUD

+++ partials/bin/bindelete.php
@@ -1,56 +1,36 @@
 <?php
     require_once 'includes/database.php';
     require_once 'includes/navbar.php';
-    $id = 0;
-     
-    if ( !empty($_GET['id'])) {
-        $id = $_REQUEST['id'];
-    }
-     
-    if ( !empty($_POST)) {
-        // keep track post values
-        $id = $_POST['id'];
-         
+  
+  if ( !empty($_POST['id']) && isset($_POST['id'])) {
+    // keep track post values
+    
+    
+    try{
         // delete data
         $pdo = Database::connect();
+        $id = $_POST['id'];
         $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-        $sql = "DELETE FROM bin  WHERE id = ?";
+        $sql = "DELETE FROM product_bin WHERE bin_id = ?";
         $q = $pdo->prepare($sql);
         $q->execute(array($id));
+        $sql = "DELETE FROM bin WHERE id = ?";
+        $q = $pdo->prepare($sql);
+        $q->execute(array($id));
+        
         Database::disconnect();
         header("Location: index.php");
-         
+    } catch (PDOException $e){
+        Database::disconnect();
+        echo $e->getMessage();
+        die();
     }
-?>
- 
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css">       <script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/js/bootstrap.min.js"></script>
-</head>
- 
-<body>
-    <div class="container">
+} else {
+    echo "failed.";
+    die();
+}
      
-                <div class="span10 offset1">
-                    <div class="row">
-                        <h3>Delete Bin</h3>
-                    </div>
-                     
-                    <form class="form-horizontal" action="delete.php" method="post">
-                      <input type="hidden" name="id" value="<?php echo $id;?>"/>
-                      <p class="alert alert-error">Are you sure to delete ?</p>
-                      <div class="form-actions">
-                          <button type="submit" class="btn btn-danger">Yes</button>
-                          <a class="btn" href="index.php">No</a>
-                        </div>
-                    </form>
-                </div>
-                 
-    </div> <!-- /container -->
-  </body>
-</html>
-<?php
+    
+
 require_once 'includes/footer.php';
 ?>
